Fix invalid code address bug using lock-based kthreads
Originally, some khttpd may be still active after unloading
khttpd.ko because we did not wait for thread exit in http_server_daemon.
The problem is solved by maitining a list shared by daemon and worker.
If the worker exits by itself, then it deletes itself from the list.
Otherwise, if the worker is still alive when unloading the module,
then the daemon sends a signal to terminate it. Besides, there is an
atomic counter to make sure module get unloaded only after all of
the worker threads exit.

Note: using kthread_stop() to an already-exited thread may corrupt the
kernel.